Sistema generador de documentos V&C/requirements.txt
pandas
openpyxl
python-docx
docxtpl
customtkinter

# Packaging / runtime extras
pyinstaller
reportlab
Pillow
pyxlsb
pymupdf
